requirements.txt
discord.py>=2.3.0
aiomysql
python-dotenv
pytz
requests
cryptography
PyNaCl
